+++ pkg/narrator/service_ai_essay_test.go
@@ -79,7 +79,7 @@ topics:
 			mockSim := &MockSim{Telemetry: sim.Telemetry{Latitude: 48.0, Longitude: 2.0}}
 
 			// Inject Essay Handler
-			svc := NewAIService(cfg, mockLLM, mockTTS, pm, mockAudio, &MockPOIProvider{}, &MockBeacon{}, mockGeo, mockSim, &MockStore{}, &MockWikipedia{}, nil, eh, nil, nil, nil)
+			svc := NewAIService(cfg, mockLLM, mockTTS, pm, mockAudio, &MockPOIProvider{}, &MockBeacon{}, mockGeo, mockSim, &MockStore{}, &MockWikipedia{}, nil, nil, eh, nil, nil, nil)
 			svc.Start()
 
 			// Action
@@ -104,7 +104,7 @@ topics:
 
 func TestAIService_PlayEssay_NoHandler(t *testing.T) {
 	// Setup minimalist service without essay handler
-	svc := NewAIService(&config.Config{}, &MockLLM{}, &MockTTS{}, nil, &MockAudio{}, &MockPOIProvider{}, &MockBeacon{}, &MockGeo{}, &MockSim{}, &MockStore{}, &MockWikipedia{}, nil, nil, nil, nil, nil)
+	svc := NewAIService(&config.Config{}, &MockLLM{}, &MockTTS{}, nil, &MockAudio{}, &MockPOIProvider{}, &MockBeacon{}, &MockGeo{}, &MockSim{}, &MockStore{}, &MockWikipedia{}, nil, nil, nil, nil, nil, nil)
 
 	if svc.PlayEssay(context.Background(), &sim.Telemetry{}) {
 		t.Error("Expected PlayEssay to return false when handler is nil")